+++ studio/src-tauri/src/preflight.rs
@@ -1,151 +1,26 @@
-use serde::{Deserialize, Serialize};
+mod backend;
+mod managed;
+mod types;
+mod version;
+
+use backend::probe_existing_backends;
+pub use managed::managed_install_ready;
+use managed::probe_managed_install;
 use std::path::PathBuf;
-use std::process::Stdio;
-use std::time::Duration;
-use tokio::io::AsyncReadExt;
-use tokio::process::Command;
+use types::{BackendProbe, ManagedProbe};
+pub use types::{DesktopPreflightDisposition, DesktopPreflightResult, ExternalBackendConflict};
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum DesktopPreflightDisposition {
-    NotInstalled,
-    ManagedReady,
-    ManagedStale,
-    AttachedReady,
-    ExternalConflict,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct DesktopPreflightResult {
-    pub disposition: DesktopPreflightDisposition,
-    pub reason: Option<String>,
-    pub port: Option<u16>,
-    pub can_auto_repair: bool,
-    pub managed_bin: Option<PathBuf>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ExternalBackendConflict {
-    pub port: u16,
-    pub reason: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum ManagedProbe {
-    Missing,
-    Ready { bin: PathBuf },
-    Stale { bin: PathBuf, reason: String },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum BackendProbe {
-    Missing,
-    Ready { port: u16 },
-    Old { port: u16, reason: String },
-    ExternalConflict { port: u16, reason: String },
-}
-
-#[derive(Debug, Deserialize)]
-struct DesktopCapability {
-    desktop_protocol_version: Option<u16>,
-    desktop_manageability_version: Option<u16>,
-    supports_api_only: Option<bool>,
-    supports_provision_desktop_auth: Option<bool>,
-    supports_desktop_backend_ownership: Option<bool>,
-    desktop_auth_stale_reason: Option<String>,
-    version: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct DesktopOwnerHealth {
-    kind: Option<String>,
-    token_sha256: Option<String>,
-}
-
-#[derive(Debug)]
-struct BackendHealth {
-    desktop_protocol_version: Option<u16>,
-    desktop_manageability_version: Option<u16>,
-    supports_desktop_auth: Option<bool>,
-    supports_desktop_backend_ownership: Option<bool>,
-    studio_root_id: Option<String>,
-    desktop_owner: Option<DesktopOwnerHealth>,
-    version: Option<String>,
-    stale_reason: Option<String>,
-}
-
-const DESKTOP_PROTOCOL_VERSION: u16 = 1;
-const DESKTOP_MANAGEABILITY_VERSION: u16 = 1;
-// Explicit backend package minimum, not the desktop app Cargo version: backend
-// and app releases can diverge. When bumping, verify this package exists on PyPI.
-const MIN_DESKTOP_BACKEND_VERSION: &str = "2026.5.2";
+#[cfg(test)]
+use backend::{backend_desktop_auth_status, backend_health};
+#[cfg(test)]
+use managed::probe_managed_bin;
+#[cfg(test)]
+use version::{
+    backend_version_compatible, backend_version_stale_reason, MIN_DESKTOP_BACKEND_VERSION,
+};
 
 fn release_auto_repair() -> bool {
     !cfg!(debug_assertions)
-}
-
-fn parse_version(value: &str) -> Option<[u64; 3]> {
-    let value = value.trim();
-    let mut parts = value.splitn(3, '.');
-    let major = parts.next()?.parse().ok()?;
-    let minor = parts.next()?.parse().ok()?;
-    let patch_and_suffix = parts.next()?;
-    let patch_len = patch_and_suffix
-        .find(|c: char| !c.is_ascii_digit())
-        .unwrap_or(patch_and_suffix.len());
-    if patch_len == 0 {
-        return None;
-    }
-    let suffix = &patch_and_suffix[patch_len..];
-    if !version_suffix_allowed(suffix) {
-        return None;
-    }
-    Some([major, minor, patch_and_suffix[..patch_len].parse().ok()?])
-}
-
-fn version_suffix_allowed(suffix: &str) -> bool {
-    suffix.is_empty()
-        || suffix.starts_with('+')
-        || suffix.starts_with(".post")
-        || suffix.starts_with(".dev")
-        || suffix.starts_with(".rc")
-        || suffix.starts_with("post")
-        || suffix.starts_with("dev")
-        || suffix.starts_with("rc")
-        || suffix.starts_with('a')
-        || suffix.starts_with('b')
-}
-
-fn backend_version_compatible(version: Option<&str>) -> bool {
-    let Some(version) = version else {
-        return false;
-    };
-    if cfg!(debug_assertions) && version == "dev" {
-        return true;
-    }
-    let Some(actual) = parse_version(version) else {
-        return false;
-    };
-    let Some(minimum) = parse_version(MIN_DESKTOP_BACKEND_VERSION) else {
-        return false;
-    };
-    actual >= minimum
-}
-
-fn backend_version_stale_reason(version: Option<&str>) -> Option<String> {
-    if backend_version_compatible(version) {
-        return None;
-    }
-    match version {
-        None | Some("") => Some("desktop_backend_version_missing".to_string()),
-        Some("dev") if !cfg!(debug_assertions) => {
-            Some("desktop_backend_version_invalid".to_string())
-        }
-        Some(value) if parse_version(value).is_none() => {
-            Some("desktop_backend_version_invalid".to_string())
-        }
-        Some(_) => Some("desktop_backend_version_too_old".to_string()),
-    }
 }
 
 fn managed_bin_for_result(managed: &ManagedProbe) -> Option<PathBuf> {
@@ -195,417 +70,6 @@ fn choose_preflight(managed: ManagedProbe, backend: BackendProbe) -> DesktopPref
             },
         },
     }
-}
-
-async fn run_cli_probe(bin: &std::path::Path, args: &[&str]) -> bool {
-    let mut cmd = Command::new(bin);
-    cmd.args(args).stdout(Stdio::null()).stderr(Stdio::null());
-
-    #[cfg(target_os = "linux")]
-    if std::env::var_os("APPIMAGE").is_some() {
-        cmd.env_remove("LD_LIBRARY_PATH");
-        cmd.env_remove("PYTHONHOME");
-        cmd.env_remove("PYTHONPATH");
-    }
-
-    // Tauri uses the legacy root regardless of UNSLOTH_STUDIO_HOME / STUDIO_HOME;
-    // probe subprocesses must follow the same isolation as process.rs.
-    cmd.env_remove("UNSLOTH_STUDIO_HOME");
-    cmd.env_remove("STUDIO_HOME");
-
-    #[cfg(windows)]
-    {
-        use std::os::windows::process::CommandExt;
-        cmd.creation_flags(crate::process::CREATE_NO_WINDOW);
-    }
-
-    let Ok(mut child) = cmd.spawn() else {
-        return false;
-    };
-
-    match tokio::time::timeout(Duration::from_secs(10), child.wait()).await {
-        Ok(Ok(status)) => status.success(),
-        _ => {
-            let _ = child.kill().await;
-            let _ = child.wait().await;
-            false
-        }
-    }
-}
-
-async fn probe_cli_capability(bin: &std::path::Path) -> Option<DesktopCapability> {
-    let mut cmd = Command::new(bin);
-    cmd.args(["studio", "desktop-capabilities", "--json"])
-        .stdout(Stdio::piped())
-        .stderr(Stdio::null());
-
-    #[cfg(target_os = "linux")]
-    if std::env::var_os("APPIMAGE").is_some() {
-        cmd.env_remove("LD_LIBRARY_PATH");
-        cmd.env_remove("PYTHONHOME");
-        cmd.env_remove("PYTHONPATH");
-    }
-
-    // Tauri uses the legacy root regardless of UNSLOTH_STUDIO_HOME / STUDIO_HOME;
-    // probe subprocesses must follow the same isolation as process.rs.
-    cmd.env_remove("UNSLOTH_STUDIO_HOME");
-    cmd.env_remove("STUDIO_HOME");
-
-    #[cfg(windows)]
-    {
-        use std::os::windows::process::CommandExt;
-        cmd.creation_flags(crate::process::CREATE_NO_WINDOW);
-    }
-
-    let Ok(mut child) = cmd.spawn() else {
-        return None;
-    };
-    let Some(mut stdout) = child.stdout.take() else {
-        return None;
-    };
-
-    match tokio::time::timeout(Duration::from_secs(10), child.wait()).await {
-        Ok(Ok(status)) if status.success() => {}
-        Err(_) => {
-            let _ = child.kill().await;
-            let _ = child.wait().await;
-            return None;
-        }
-        _ => return None,
-    }
-
-    let mut output = Vec::new();
-    if stdout.read_to_end(&mut output).await.is_err() {
-        return None;
-    }
-
-    serde_json::from_slice::<DesktopCapability>(&output).ok()
-}
-
-fn desktop_capability_stale_reason(capability: &DesktopCapability) -> Option<String> {
-    if capability.desktop_protocol_version != Some(DESKTOP_PROTOCOL_VERSION) {
-        return Some("desktop_protocol_incompatible".to_string());
-    }
-    if capability.supports_api_only != Some(true) {
-        return Some("desktop_api_only_unsupported".to_string());
-    }
-    if capability.supports_provision_desktop_auth != Some(true) {
-        return capability
-            .desktop_auth_stale_reason
-            .clone()
-            .or_else(|| Some("desktop_auth_unsupported".to_string()));
-    }
-    if capability.desktop_manageability_version.unwrap_or(0) < DESKTOP_MANAGEABILITY_VERSION {
-        return Some("desktop_manageability_unsupported".to_string());
-    }
-    if capability.supports_desktop_backend_ownership != Some(true) {
-        return Some("desktop_backend_ownership_unsupported".to_string());
-    }
-    backend_version_stale_reason(capability.version.as_deref())
-}
-
-fn desktop_capability_ready(capability: &DesktopCapability) -> bool {
-    desktop_capability_stale_reason(capability).is_none()
-}
-
-async fn probe_managed_bin(bin: PathBuf) -> ManagedProbe {
-    if !run_cli_probe(&bin, &["-h"]).await {
-        return ManagedProbe::Stale {
-            bin,
-            reason: "cli_unusable".to_string(),
-        };
-    }
-
-    let capability = probe_cli_capability(&bin).await;
-    if let Some(capability) = capability {
-        if desktop_capability_ready(&capability) {
-            return ManagedProbe::Ready { bin };
-        }
-        return ManagedProbe::Stale {
-            bin,
-            reason: desktop_capability_stale_reason(&capability)
-                .unwrap_or_else(|| "desktop_capability_incompatible".to_string()),
-        };
-    }
-
-    ManagedProbe::Stale {
-        bin,
-        reason: "desktop_capability_probe_failed".to_string(),
-    }
-}
-
-async fn probe_managed_install() -> ManagedProbe {
-    match crate::process::find_unsloth_binary() {
-        Some(bin) => probe_managed_bin(bin).await,
-        None => ManagedProbe::Missing,
-    }
-}
-
-pub async fn managed_install_ready() -> bool {
-    matches!(probe_managed_install().await, ManagedProbe::Ready { .. })
-}
-
-async fn backend_health(client: &reqwest::Client, port: u16) -> Option<BackendHealth> {
-    let url = format!("http://127.0.0.1:{port}/api/health");
-    let response = client.get(url).send().await.ok()?;
-    if !response.status().is_success() {
-        return None;
-    }
-    let json = response.json::<serde_json::Value>().await.ok()?;
-    let healthy = json
-        .get("status")
-        .and_then(|v| v.as_str())
-        .map(|s| s == "healthy")
-        .unwrap_or(false);
-    let service = json
-        .get("service")
-        .and_then(|v| v.as_str())
-        .map(|s| s == "Unsloth UI Backend")
-        .unwrap_or(false);
-    if !healthy || !service {
-        return None;
-    }
-
-    let desktop_protocol_version = json
-        .get("desktop_protocol_version")
-        .and_then(|v| v.as_u64())
-        .and_then(|v| u16::try_from(v).ok());
-    let desktop_manageability_version = json
-        .get("desktop_manageability_version")
-        .and_then(|v| v.as_u64())
-        .and_then(|v| u16::try_from(v).ok());
-    let supports_desktop_auth = json.get("supports_desktop_auth").and_then(|v| v.as_bool());
-    let supports_desktop_backend_ownership = json
-        .get("supports_desktop_backend_ownership")
-        .and_then(|v| v.as_bool());
-    let studio_root_id = json
-        .get("studio_root_id")
-        .and_then(|v| v.as_str())
-        .map(ToOwned::to_owned);
-    let desktop_owner = json
-        .get("desktop_owner")
-        .and_then(|v| serde_json::from_value::<DesktopOwnerHealth>(v.clone()).ok());
-    let version = json
-        .get("version")
-        .and_then(|v| v.as_str())
-        .map(ToOwned::to_owned);
-    let stale_reason = match supports_desktop_auth {
-        Some(false) => json
-            .get("desktop_auth_stale_reason")
-            .and_then(|v| v.as_str())
-            .map(ToOwned::to_owned),
-        _ => None,
-    };
-    Some(BackendHealth {
-        desktop_protocol_version,
-        desktop_manageability_version,
-        supports_desktop_auth,
-        supports_desktop_backend_ownership,
-        studio_root_id,
-        desktop_owner,
-        version,
-        stale_reason,
-    })
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum BackendRootStatus {
-    SameRoot,
-    ForeignRoot,
-    AmbiguousRoot,
-    ExpectedUnavailable,
-}
-
-fn backend_root_status(
-    health: &BackendHealth,
-    expected_studio_root_id: Option<&str>,
-) -> BackendRootStatus {
-    let Some(expected) = expected_studio_root_id else {
-        return BackendRootStatus::ExpectedUnavailable;
-    };
-
-    match health.studio_root_id.as_deref() {
-        Some(actual) if actual == expected => BackendRootStatus::SameRoot,
-        Some(actual) if crate::desktop_backend_owner::is_valid_studio_root_id(actual) => {
-            BackendRootStatus::ForeignRoot
-        }
-        _ => BackendRootStatus::AmbiguousRoot,
-    }
-}
-
-fn backend_has_verified_desktop_owner(health: &BackendHealth) -> bool {
-    let Some(owner) = health.desktop_owner.as_ref() else {
-        return false;
-    };
-    crate::desktop_backend_owner::health_matches_desktop_owner(
-        health.studio_root_id.as_deref(),
-        owner.kind.as_deref(),
-        owner.token_sha256.as_deref(),
-    )
-}
-
-fn backend_capability_stale_reason(health: &BackendHealth) -> Option<String> {
-    if health.desktop_protocol_version != Some(DESKTOP_PROTOCOL_VERSION) {
-        return health
-            .stale_reason
-            .clone()
-            .or_else(|| Some("desktop_protocol_incompatible".to_string()));
-    }
-    if health.supports_desktop_auth != Some(true) {
-        return health
-            .stale_reason
-            .clone()
-            .or_else(|| Some("desktop_auth_unsupported".to_string()));
-    }
-    if health.desktop_manageability_version.unwrap_or(0) < DESKTOP_MANAGEABILITY_VERSION {
-        return Some("desktop_manageability_unsupported".to_string());
-    }
-    if health.supports_desktop_backend_ownership != Some(true) {
-        return Some("desktop_backend_ownership_unsupported".to_string());
-    }
-    backend_version_stale_reason(health.version.as_deref())
-}
-
-#[derive(Serialize)]
-struct DesktopLoginProbe<'a> {
-    secret: &'a str,
-}
-
-async fn backend_desktop_auth_status(
-    client: &reqwest::Client,
-    port: u16,
-    health: &BackendHealth,
-    expected_studio_root_id: Option<&str>,
-) -> BackendProbe {
-    let root_status = backend_root_status(health, expected_studio_root_id);
-    let verified_owner = backend_has_verified_desktop_owner(health);
-    let same_root_external = root_status == BackendRootStatus::SameRoot && !verified_owner;
-
-    match root_status {
-        BackendRootStatus::AmbiguousRoot | BackendRootStatus::ExpectedUnavailable => {
-            return BackendProbe::ExternalConflict {
-                port,
-                reason: "ambiguous_root_external_backend_active".to_string(),
-            };
-        }
-        BackendRootStatus::ForeignRoot => {
-            return BackendProbe::Old {
-                port,
-                reason: "studio_root_id_mismatch".to_string(),
-            };
-        }
-        BackendRootStatus::SameRoot => {}
-    }
-
-    if let Some(reason) = backend_capability_stale_reason(health) {
-        return if same_root_external {
-            BackendProbe::ExternalConflict { port, reason }
-        } else {
-            BackendProbe::Old { port, reason }
-        };
-    }
-
-    let url = format!("http://127.0.0.1:{port}/api/auth/desktop-login");
-    let response = client
-        .post(url)
-        .json(&DesktopLoginProbe {
-            secret: "desktop-preflight-invalid-secret",
-        })
-        .send()
-        .await;
-
-    let Ok(response) = response else {
-        let reason = backend_capability_stale_reason(health)
-            .unwrap_or_else(|| "desktop_login_probe_failed".to_string());
-        return if same_root_external {
-            BackendProbe::ExternalConflict { port, reason }
-        } else {
-            BackendProbe::Old { port, reason }
-        };
-    };
-
-    match response.status() {
-        reqwest::StatusCode::UNAUTHORIZED => BackendProbe::Ready { port },
-        reqwest::StatusCode::NOT_FOUND => {
-            let reason = "desktop_login_not_found".to_string();
-            if same_root_external {
-                BackendProbe::ExternalConflict { port, reason }
-            } else {
-                BackendProbe::Old { port, reason }
-            }
-        }
-        _ => {
-            let reason = backend_capability_stale_reason(health)
-                .unwrap_or_else(|| "desktop_login_probe_failed".to_string());
-            if same_root_external {
-                BackendProbe::ExternalConflict { port, reason }
-            } else {
-                BackendProbe::Old { port, reason }
-            }
-        }
-    }
-}
-
-async fn probe_existing_backends(ignored_ports: &[u16]) -> BackendProbe {
-    let client = match reqwest::Client::builder()
-        .timeout(Duration::from_secs(2))
-        .build()
-    {
-        Ok(client) => client,
-        Err(_) => return BackendProbe::Missing,
-    };
-
-    // Fan out health probes concurrently. The desktop-auth probe is still
-    // sequential per candidate because it has auth-log side effects.
-    let ports: Vec<u16> = (8888u16..=8908).collect();
-    let mut health_futs = Vec::with_capacity(ports.len());
-    for port in ports {
-        // why: reqwest::Client is internally Arc-wrapped; clone is a refcount bump
-        // (documented cheap). tokio::spawn needs 'static, so each task owns its own clone.
-        let c = client.clone();
-        health_futs.push(tokio::spawn(async move {
-            backend_health(&c, port).await.map(|h| (port, h))
-        }));
-    }
-
-    let mut candidates: Vec<(u16, BackendHealth)> = Vec::new();
-    for fut in health_futs {
-        if let Ok(Some(pair)) = fut.await {
-            candidates.push(pair);
-        }
-    }
-
-    let expected_studio_root_id = crate::desktop_backend_owner::read_expected_studio_root_id();
-    let mut first_conflict = None;
-    let mut first_ready = None;
-    let mut first_old = None;
-    for (port, health) in candidates {
-        if ignored_ports.contains(&port) {
-            continue;
-        }
-        match backend_desktop_auth_status(
-            &client,
-            port,
-            &health,
-            expected_studio_root_id.as_deref(),
-        )
-        .await
-        {
-            conflict @ BackendProbe::ExternalConflict { .. } if first_conflict.is_none() => {
-                first_conflict = Some(conflict)
-            }
-            ready @ BackendProbe::Ready { .. } if first_ready.is_none() => {
-                first_ready = Some(ready)
-            }
-            old @ BackendProbe::Old { .. } if first_old.is_none() => first_old = Some(old),
-            _ => {}
-        }
-    }
-
-    first_conflict
-        .or(first_ready)
-        .or(first_old)
-        .unwrap_or(BackendProbe::Missing)
 }
 
 fn mutation_blocker_from_probe(probe: BackendProbe) -> Option<ExternalBackendConflict> {
